状態を一つに統合
